CLOUD-333: moving split function for separation of concern
